Create useNavigation. Move some states to it.
